Organized hierarchy, added packages and settings
I organized the hierarchy so that it's clear what objects are doing what.  I also added the various packages needed for networking, as per the tutorial from class.  Lastly, I also fixed the resolution settings so that the window is resizeable and windowed by default.  The project is linked to my Unity dashboard, but I don't know if it'll stay that way, so we'll have to see what happens with that.  I found a place I can probably add people to the unity project, but I can't confirm if it works or not yet.