Inventory Text
Just added the inventory text on the side of the screen.  It does not currently work, as I can't seem to add the "Inventory" object (TextMeshPro) into the slot I made for it in the inspector panel of the PlayerMovement script.  This part is still a work in progress, so it shouldn't work.  For that reason, you can disable it or try and figure out why it's not working, because I have no clue right now.  Also, I added a way to see how many parts you currently have in the inventory in the inspector called "Current Part Count".  I hope this helps!