Reworked Inventory System + More
I added quite a few things here.  I'll start by saying I added a check for making sure that when an item is added to the inventory, it hasn't already been turned in yet.  I haven't handled what happens if you get two of the same item in the inventory since the way I'm using it for right now makes that impossible.  I also enforced the maxParts for how many parts can be in the player's inventory at a given time.  Previously, using the 'I' key would add a part to your inventory; this has been changed now so that it always gives you the next item in the parts list, which avoids duplicates.  Speaking of, there is now a general parts list that holds the actual Part objects being used in the inventory.  I made a note about how to use that in the AddToInventory function, so please give that a read.  Long story short, all parts are stored in one list and adding/subtracting to the inventory directly interacts with them instead of making local copies.  Pressing the 'R' key now provides a way for you to reset the "wasTurnedIn" status of all items, allowing you to keep testing more easily.  I also added some utility functions, like initParts and findPartIndexByName, which both do what they sound like; initialize the parts list with all the parts and added a way to easily search for a part name in the parts list and return it's index to you.  That reminds me, I also added a "memberOf" variable to the Part class, which essentially holds which category it's in; "Input", "CPU", or "Output" devices, as shown in the diagram given to us in the project documentation.  I don't think there's anything else, but let me know if you need any help!