Prevents Duplicate Items in Inventory
I added a simple function that can check if a given name is already in the inventory list.  So now, when adding to the inventory, it confirms that the item does not already exist in the inventory list; in other words, this should prevent duplicate items from entering the inventory.  Also, I increased the maxBullet count from 5 to 15 so that testing can go a bit faster.