Added UI
I added a UI element that holds all items to the screen, and provided a function to update valid parts.  That said, I haven't been able to get this to work since I can't seem to drag and drop the inventoryManager object into the playerMovement script.  Might be able to fix that by making a prefab, so might test that next.