Updated UI - Visually Works Now
Alright, so I fixed the UI so now when you convert a part, it correctly updates the visual list by updating the top most item that wasn't already turned in.  However, there seems to be an issue where if you have, for example, 3 parts and turn them all in, it still lets you convert them in to bullets.  This leads to some weird errors, like having a negative part count.  Still need to look into this, but at least it sorta works.