Updated UI + Conversion/TurnIn
Alright, so hopefully this fixed a few different issues.  The main thing is that you can now only convert parts into bullets as long as you actually have parts to convert.  This was causing issues because turning in items wasn't properly adjusting this.  However, a new issue is that if you turn in a bunch of items, then convert a couple, next time you try to get an item, you won't be able to since it was marked as already turned in.  In order to fix this, you have to reset all the values by pressing 'R' and re-giving yourself all the items and turning them all in again.  Obviously, this is not intended, but at least for now, the basic of basic functionality is there.